Modifications, regarding Regression run preparation - July 03 . 2022
