+++ LingaAllure/src/main/java/LingaAndroid/LingaAllure/Utilities.java
@@ -1,43 +1,37 @@
 package LingaAndroid.LingaAllure;
 
-import java.time.Duration;
+import java.io.IOException;
+//import java.time.Duration;
 import java.util.logging.Logger;
-
-import org.openqa.selenium.Dimension;
-import org.openqa.selenium.Rectangle;
-import org.openqa.selenium.WebElement;
+//import org.openqa.selenium.Dimension;
+//import org.openqa.selenium.Rectangle;
+//import org.openqa.selenium.WebElement;
 import org.testng.Assert;
-
-
-
 import Linga.pageObjects.CheckStatusPage;
+import Linga.pageObjects.OrderType;
 import Linga.pageObjects.PinPage;
 import Linga.pageObjects.StoreLoginPage;
-import io.appium.java_client.AppiumDriver;
-import io.appium.java_client.MobileElement;
-import io.appium.java_client.TouchAction;
+//import io.appium.java_client.AppiumDriver;
+//import io.appium.java_client.MobileElement;
+//import io.appium.java_client.TouchAction;
 import io.appium.java_client.android.AndroidDriver;
 import io.appium.java_client.android.AndroidElement;
 import io.appium.java_client.android.connection.ConnectionStateBuilder;
-import io.appium.java_client.touch.WaitOptions;
-import io.appium.java_client.touch.offset.PointOption;
+//import io.appium.java_client.touch.WaitOptions;
+//import io.appium.java_client.touch.offset.PointOption;
 
 
-
-
-public class Utilities {
+public class Utilities extends LingaBase{
 
 	AndroidDriver<AndroidElement> dr;
 
-	public Utilities(AndroidDriver<AndroidElement> dr)
-	{
+	public Utilities(AndroidDriver<AndroidElement> dr) {
+		
 		this.dr=dr;
 	}
 
 
-
-	public void storeLogin(String email, String password)
-	{
+	public void storeLogin(String email, String password) {
 
 		StoreLoginPage slp = new StoreLoginPage(dr);
 		slp.EmailAddress.sendKeys(email);
@@ -46,29 +40,35 @@ public class Utilities {
 
 	}
 
-	public void defaultPin() throws InterruptedException
-	{
+	public void defaultPin() throws InterruptedException  {
 
 		PinPage pp = new PinPage(dr);
 		pp.One.click();
 		pp.Two.click();
-		//Thread.sleep(2000);
 		pp.Three.click();
 		pp.Four.click();
-		//pp.Backspace.click();
-		//pp.Four.click();
 		pp.Login.click();
-		//Thread.sleep(4000);
+	}
+
+	public void DashboardModuleVerify() throws InterruptedException, IOException {
+
+		try {
+			if(dr.findElementByXPath("//*[@text='Order Type']").isDisplayed())
+			{
+				OrderType otp = new OrderType(dr);
+				otp.Cancel.click();
+			}
+		}
+		catch (Exception e) {
+			InstallAndroid();
+		}
 	}
 
 	public void verifyLoginUser() {
 
 		CheckStatusPage csp = new CheckStatusPage(dr);
-
 		String str =  csp.PosUser.getText();
 		Assert.assertEquals("shri admin 1", str);
-
-
 	}
 
 
@@ -86,14 +86,10 @@ public class Utilities {
 	public void setConnectionToON() {
 		try {
 			dr.setConnection(new ConnectionStateBuilder().withWiFiEnabled().build());
-
 			logger.info("Switching On the connection: " + dr.getConnection());
 		} catch (Exception e) {
 			logger.info("Connection could not be switch ON");
 		}
 	}
-
-
-
 
 }
