+++ LingaAllure/src/main/java/global.properties
@@ -1,3 +1,3 @@
-apprelease=pos5.5.42.apk
+apprelease=pos5.5.53.apk
 cds=Cds303121.apk
 device=Android Device
