+++ LingaAllure/src/test/java/LingaRos/LingaPos/TestCaseTwo.java
@@ -1,0 +1,21 @@
+package LingaRos.LingaPos;
+
+import java.io.IOException;
+import org.testng.annotations.Test;
+import LingaAndroid.LingaAllure.LingaBase;
+import LingaAndroid.LingaAllure.Utilities;
+import io.appium.java_client.android.AndroidDriver;
+import io.appium.java_client.android.AndroidElement;
+
+public class TestCaseTwo {
+
+
+	
+	
+	@Test
+	public void PlaceOrder() throws InterruptedException, IOException {
+
+	
+	}
+
+}
