+++ LingaAllure/test-output/Default suite/Default test.html
@@ -57,9 +57,9 @@ function toggleAllBoxes() {
 <tr>
 <td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Sat Jun 04 17:49:27 IST 2022</td>
+<td>Started on:</td><td>Sun Jul 03 16:21:59 IST 2022</td>
 </tr>
-<tr><td>Total time:</td><td>44 seconds (44970 ms)</td>
+<tr><td>Total time:</td><td>7 seconds (7902 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -75,10 +75,10 @@ function toggleAllBoxes() {
 <td><b>Instance</b></td>
 </tr>
 <tr>
-<td title='LingaAndroid.LingaAllure.RandomTest.tryOne()'><b>tryOne</b><br>Test class: LingaAndroid.LingaAllure.RandomTest</td>
+<td title='LingaRos.LingaPos.TestCaseOne.PlaceOrder()'><b>PlaceOrder</b><br>Test class: LingaRos.LingaPos.TestCaseOne</td>
 <td></td>
-<td>44</td>
-<td>LingaAndroid.LingaAllure.RandomTest@8e24743</td></tr>
+<td>7</td>
+<td>LingaRos.LingaPos.TestCaseOne@8e24743</td></tr>
 </table><p>
 </body>
 </html>
