+++ LingaAllure/test-output/Default suite/Default test.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite hostname="LAPTOP-FF5I964F" ignored="0" name="Default test" tests="1" failures="0" timestamp="2022-06-04T17:50:12 IST" time="44.97" errors="0">
-  <testcase name="tryOne" time="44.935" classname="LingaAndroid.LingaAllure.RandomTest"/>
+<testsuite hostname="LAPTOP-FF5I964F" ignored="0" name="Default test" tests="1" failures="0" timestamp="2022-07-03T16:22:07 IST" time="7.902" errors="0">
+  <testcase name="PlaceOrder" time="7.863" classname="LingaRos.LingaPos.TestCaseOne"/>
 </testsuite> <!-- Default test -->
