+++ LingaAllure/test-output/index.html
@@ -51,7 +51,7 @@
               <ul>
                 <li>
                   <a href="#" class="navigator-link " panel-name="test-xml-Default_suite">
-                    <span>C:\Users\91948\AppData\Local\Temp\testng-eclipse-769045644\testng-customsuite.xml</span>
+                    <span>C:\Users\91948\AppData\Local\Temp\testng-eclipse-866525769\testng-customsuite.xml</span>
                   </a> <!-- navigator-link  -->
                 </li>
                 <li>
@@ -104,7 +104,7 @@
                     <div class="method-list-content passed suite-Default_suite">
                       <span>
                         <img width="3%" src="passed.png"/>
-                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="LingaAndroid.LingaAllure.RandomTest" hash-for-method="tryOne">tryOne</a> <!-- method navigator-link -->
+                        <a href="#" class="method navigator-link" panel-name="suite-Default_suite" title="LingaRos.LingaPos.TestCaseOne" hash-for-method="PlaceOrder">PlaceOrder</a> <!-- method navigator-link -->
                       </span>
                       <br/>
                     </div> <!-- method-list-content passed suite-Default_suite -->
@@ -122,14 +122,14 @@
           <div class="suite-Default_suite-class-passed">
             <div class="main-panel-header rounded-window-top">
               <img src="passed.png"/>
-              <span class="class-name">LingaAndroid.LingaAllure.RandomTest</span>
+              <span class="class-name">LingaRos.LingaPos.TestCaseOne</span>
             </div> <!-- main-panel-header rounded-window-top -->
             <div class="main-panel-content rounded-window-bottom">
               <div class="method">
                 <div class="method-content">
-                  <a name="tryOne">
-                  </a> <!-- tryOne -->
-                  <span class="method-name">tryOne</span>
+                  <a name="PlaceOrder">
+                  </a> <!-- PlaceOrder -->
+                  <span class="method-name">PlaceOrder</span>
                 </div> <!-- method-content -->
               </div> <!-- method -->
             </div> <!-- main-panel-content rounded-window-bottom -->
@@ -137,7 +137,7 @@
         </div> <!-- panel Default_suite -->
         <div panel-name="test-xml-Default_suite" class="panel">
           <div class="main-panel-header rounded-window-top">
-            <span class="header-content">C:\Users\91948\AppData\Local\Temp\testng-eclipse-769045644\testng-customsuite.xml</span>
+            <span class="header-content">C:\Users\91948\AppData\Local\Temp\testng-eclipse-866525769\testng-customsuite.xml</span>
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <pre>
@@ -146,7 +146,7 @@
 &lt;suite guice-stage=&quot;DEVELOPMENT&quot; name=&quot;Default suite&quot;&gt;
   &lt;test thread-count=&quot;5&quot; verbose=&quot;2&quot; name=&quot;Default test&quot;&gt;
     &lt;classes&gt;
-      &lt;class name=&quot;LingaAndroid.LingaAllure.RandomTest&quot;/&gt;
+      &lt;class name=&quot;LingaRos.LingaPos.TestCaseOne&quot;/&gt;
     &lt;/classes&gt;
   &lt;/test&gt; &lt;!-- Default test --&gt;
 &lt;/suite&gt; &lt;!-- Default suite --&gt;
@@ -188,14 +188,14 @@ data.addColumn('string', 'Class');
 data.addColumn('number', 'Time (ms)');
 data.addRows(1);
 data.setCell(0, 0, 0)
-data.setCell(0, 1, 'tryOne')
-data.setCell(0, 2, 'LingaAndroid.LingaAllure.RandomTest')
-data.setCell(0, 3, 44935);
+data.setCell(0, 1, 'PlaceOrder')
+data.setCell(0, 2, 'LingaRos.LingaPos.TestCaseOne')
+data.setCell(0, 3, 7863);
 window.suiteTableData['Default_suite']= { tableData: data, tableDiv: 'times-div-Default_suite'}
 return data;
 }
               </script>
-              <span class="suite-total-time">Total running time: 44 seconds</span>
+              <span class="suite-total-time">Total running time: 7 seconds</span>
               <div id="times-div-Default_suite">
               </div> <!-- times-div-Default_suite -->
             </div> <!-- times-div -->
@@ -221,9 +221,9 @@ return data;
           </div> <!-- main-panel-header rounded-window-top -->
           <div class="main-panel-content rounded-window-bottom">
             <div class="chronological-class">
-              <div class="chronological-class-name">LingaAndroid.LingaAllure.RandomTest</div> <!-- chronological-class-name -->
+              <div class="chronological-class-name">LingaRos.LingaPos.TestCaseOne</div> <!-- chronological-class-name -->
               <div class="test-method">
-                <span class="method-name">tryOne</span>
+                <span class="method-name">PlaceOrder</span>
                 <span class="method-start">0 ms</span>
               </div> <!-- test-method -->
           </div> <!-- main-panel-content rounded-window-bottom -->
