+++ LingaAllure/test-output/junitreports/TEST-LingaRos.LingaPos.AndroidInstall.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite skipped="0" hostname="LAPTOP-FF5I964F" name="LingaRos.LingaPos.AndroidInstall" tests="1" failures="0" timestamp="2022-07-02T08:45:13 IST" time="19.967" errors="0">
+  <testcase name="installAndroidalone" time="19.967" classname="LingaRos.LingaPos.AndroidInstall"/>
+  <system-out/>
+</testsuite> <!-- LingaRos.LingaPos.AndroidInstall -->
