+++ LingaAllure/test-output/junitreports/TEST-LingaRos.LingaPos.Qsr_WebTest.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite skipped="0" hostname="LAPTOP-FF5I964F" name="LingaRos.LingaPos.Qsr_WebTest" tests="1" failures="0" timestamp="2022-06-28T16:35:02 IST" time="22.514" errors="0">
+  <testcase name="tryOne" time="22.514" classname="LingaRos.LingaPos.Qsr_WebTest"/>
+  <system-out/>
+</testsuite> <!-- LingaRos.LingaPos.Qsr_WebTest -->
