+++ LingaAllure/test-output/junitreports/TEST-LingaRos.LingaPos.TestCaseOne.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite skipped="0" hostname="LAPTOP-FF5I964F" name="LingaRos.LingaPos.TestCaseOne" tests="1" failures="0" timestamp="2022-07-03T16:22:07 IST" time="7.863" errors="0">
+  <testcase name="PlaceOrder" time="7.863" classname="LingaRos.LingaPos.TestCaseOne"/>
+  <system-out/>
+</testsuite> <!-- LingaRos.LingaPos.TestCaseOne -->
