+++ LingaAllure/test-output/junitreports/TEST-TestPos.LingaRos.LingaTestInstall.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitReportReporter -->
+<testsuite skipped="0" hostname="LAPTOP-FF5I964F" name="TestPos.LingaRos.LingaTestInstall" tests="1" failures="0" timestamp="2022-06-25T16:49:48 IST" time="40.274" errors="0">
+  <testcase name="tryOne" time="40.274" classname="TestPos.LingaRos.LingaTestInstall"/>
+  <system-out/>
+</testsuite> <!-- TestPos.LingaRos.LingaTestInstall -->
