+++ LingaAllure/test-output/old/Default suite/classes.html
@@ -4,14 +4,14 @@
 <th>Method name</th>
 <th>Groups</th>
 </tr><tr>
-<td>LingaAndroid.LingaAllure.RandomTest</td>
+<td>LingaRos.LingaPos.TestCaseOne</td>
 <td>&nbsp;</td><td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@Test</td>
 </tr>
 <tr>
 <td>&nbsp;</td>
-<td>tryOne</td>
+<td>PlaceOrder</td>
 <td>&nbsp;</td></tr>
 <tr>
 <td align='center' colspan='3'>@BeforeClass</td>
