+++ LingaAllure/test-output/old/Default suite/methods.html
@@ -1,6 +1,6 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="c2a581">  <td>22/06/04 17:49:28</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="RandomTest.tryOne()[pri:0, instance:LingaAndroid.LingaAllure.RandomTest@8e24743]">tryOne</td> 
+<tr bgcolor="c2a581">  <td>22/07/03 16:21:59</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="TestCaseOne.PlaceOrder()[pri:0, instance:LingaRos.LingaPos.TestCaseOne@8e24743]">PlaceOrder</td> 
   <td>main@85445963</td>   <td></td> </tr>
 </table>
