+++ LingaAllure/test-output/testng-failed.xml
@@ -3,11 +3,11 @@
 <suite guice-stage="DEVELOPMENT" name="Failed suite [Default suite]">
   <test thread-count="5" name="Default test(failed)">
     <classes>
-      <class name="LingaAndroid.LingaAllure.RandomTest">
+      <class name="TestPos.LingaRos.Qsr_AllureTest">
         <methods>
           <include name="tryOne"/>
         </methods>
-      </class> <!-- LingaAndroid.LingaAllure.RandomTest -->
+      </class> <!-- TestPos.LingaRos.Qsr_AllureTest -->
     </classes>
   </test> <!-- Default test(failed) -->
 </suite> <!-- Failed suite [Default suite] -->
